+++ chats/apps/api/v1/projects/serializers.py
@@ -6,6 +6,8 @@ from chats.apps.projects.models import (
     LinkContact,
     Project,
     ProjectPermission,
+    CustomStatusType,
+    CustomStatus,
 )
 from chats.apps.sectors.models import Sector
 
@@ -16,7 +18,7 @@ class ProjectSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Project
-        fields = ["name", "date_format", "timezone", "config"]
+        fields = ["name", "date_format", "timezone", "config", "org"]
         read_only_fields = [
             "timezone",
         ]
@@ -103,3 +105,28 @@ class ListProjectUsersSerializer(serializers.ModelSerializer):
     class Meta:
         model = ProjectPermission
         fields = ["first_name", "last_name", "email", "photo_url"]
+
+
+class CustomStatusTypeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = CustomStatusType
+        fields = ["uuid", "name", "project", "is_deleted"]
+        read_only_fields = ["uuid", "is_deleted"]
+
+
+class CustomStatusSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = CustomStatus
+        fields = [
+            "uuid",
+            "user",
+            "status_type",
+            "is_active",
+            "break_time",
+            "created_on",
+        ]
+        read_only_fields = [
+            "uuid",
+        ]
+
+    break_time = serializers.IntegerField(required=False)