+++ chats/apps/api/v1/rooms/serializers.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 from rest_framework import serializers
 
@@ -100,6 +101,7 @@ class ListRoomSerializer(serializers.ModelSerializer):
 
     last_interaction = serializers.DateTimeField(read_only=True)
     can_edit_custom_fields = serializers.SerializerMethodField()
+    is_active = serializers.BooleanField(default=True)
 
     class Meta:
         model = Room
@@ -119,6 +121,7 @@ class ListRoomSerializer(serializers.ModelSerializer):
             "transfer_history",
             "protocol",
             "service_chat",
+            "is_active",
         ]
 
     def get_user(self, room: Room):
@@ -231,3 +234,37 @@ class RoomContactSerializer(serializers.ModelSerializer):
 
     def get_is_waiting(self, room: Room):
         return room.get_is_waiting()
+
+
+class RoomInfoSerializer(serializers.ModelSerializer):
+    first_user_message_sent_at = serializers.SerializerMethodField()
+    user = serializers.SerializerMethodField()
+
+    class Meta:
+        model = Room
+        fields = [
+            "uuid",
+            "user",
+            "first_user_message_sent_at",
+            "user_assigned_at",
+        ]
+
+    def get_user(self, room: Room) -> dict:
+        user: User = room.user
+
+        if not user:
+            return None
+
+        name = f"{user.first_name} {user.last_name}".strip()
+
+        return {"email": user.email, "name": name}
+
+    def get_first_user_message_sent_at(self, room: Room) -> datetime:
+        if (
+            first_user_message := room.messages.filter(user__isnull=False)
+            .order_by("created_on")
+            .first()
+        ):
+            return first_user_message.created_on
+
+        return None